change file name & move to types folder
