+++ packages/frinx-workflow-ui/src/pages/workflowList/WorkflowExec/DetailsModal/task-table.tsx
@@ -1,7 +1,7 @@
 // @flow
 import React, { FC } from 'react';
 import { Button, Table, Tbody, Td, Text, Th, Thead, Tooltip, Tr } from '@chakra-ui/react';
-import { Task } from '../../../../common/flowtypes';
+import { Task } from '../../../../types/task';
 
 type Props = {
   tasks: Task[];
